+++ Kahla.Server/Models/ApiAddressModels/AllAddressModel.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace Kahla.Server.Models.ApiAddressModels
 {
-    public class MyFriendsAddressModel
+    public class AllAddressModel
     {
         [Range(1, int.MaxValue)]
         public int Take { get; set; } = 15;
