+++ Kahla.Server/Models/ApiViewModels/MineViewModel.cs
@@ -1,0 +1,14 @@
+﻿using Aiursoft.Pylon.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Kahla.Server.Models.ApiViewModels
+{
+    public class MineViewModel : AiurProtocol
+    {
+        public List<KahlaUser> Users { get; internal set; }
+        public List<GroupConversation> Groups { get; internal set; }
+    }
+}
